+++ docs/changelog.rst
@@ -2,9 +2,18 @@
 Changelog
 #####################################
 
+1.0.1
+------
+
+December 05, 2022
+
+* Fix: settings can be animated
+* Fix: max. detail height isn't enough, add multiplier
+* Fix: documentation link in the panel
+
 1.0.0
 ------
 
-  November 04, 2022
+November 04, 2022
 
 * Initial release